+++ backend/package.json
@@ -7,7 +7,6 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:debug": "jest --detectOpenHandles",
-    "postinstall": "node scriptdb.js",
     "start": "node index.js",
     "dev": "nodemon index.js"
   },
